[Screenshot] Folder 'DHEMS_screenshot'
- screenshot index.html & loadFix.hmtl in web.py
- screenshot backupLHEMS & backupGHEMS in backup.py
- two mode 'RECOVER' & 'SAVEAS' in backup.py
	- RECOVER: DEL original file then re-screenshot the same img
	- SAVEAS: Create another folder or img which named include (1)
- backup.py screenshot without each household cost (LHEMS_cost)
