+++ DHEMS_screenshot/.gitignore
@@ -1,0 +1,1 @@
+utils/__pycache__
