+++ DHEMS_screenshot/utils/__init__.py
@@ -1,0 +1,150 @@
+
+class URL:
+    def __init__(self) -> None:
+        self.phpmyadmin = 'http://140.124.42.65/phpmyadmin'
+        self.DHEMS_web_loadFix = 'http://140.124.42.65/how/DHEMS_web/loadFix.html'
+        self.DHEMS_web_index = 'http://140.124.42.65/how/DHEMS_web/index.html'
+        self.DHEMS_web_baseParameter = 'http://140.124.42.65/how/DHEMS_web/baseParameter.html'
+        self.DHEMS_web_backup_LHEMS = 'http://140.124.42.65/how/DHEMS_web/backup_LHEMS.html'
+        self.DHEMS_web_backup_GHEMS = 'http://140.124.42.65/how/DHEMS_web/backup_GHEMS.html'
+
+class Xpath:
+    def __init__(self) -> None:
+        self.input_username = '//*[@id="input_username"]'
+        self.input_passowrd = '//*[@id="input_password"]'
+        self.btn_login = '//*[@id="input_go"]'
+        self.btn_DHEMS_group = '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[1]/a/img'
+        self.text_DB_DHMES_fiftyHousehold = '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/a'
+        self.alert_loading = '//*[@id="pma_navigation_header"]/img'
+        # menu bar info
+        self.serverinfo_DB = '//*[@id="serverinfo"]/a[2]'
+        self.serverinfo_table = '//*[@id="serverinfo"]/a[3]'
+        # DHEMS_fiftyHousehold tables
+        self.text_create_newTable =                 '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[2]/a'
+        self.text_backup_BaseParameter =            '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[3]/a'
+        self.text_backup_EM_Parameter =             '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[4]/a'
+        self.text_backup_EM_user_number =           '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[5]/a'
+        self.text_backup_EM_user_result =           '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[6]/a'
+        self.text_backup_EV_Parameter =             '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[7]/a'
+        self.text_backup_EV_user_number =           '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[8]/a'
+        self.text_backup_EV_user_result =           '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[9]/a'
+        self.text_backup_GHEMS =                    '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[10]/a'
+        self.text_backup_LHEMS =                    '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[11]/a'
+        self.text_backup_totalLoad =                '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[12]/a'
+        self.text_BaseParameter =                   '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[13]/a'
+        self.text_cost =                            '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[14]/a'
+        self.text_dr =                              '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[15]/a'
+        self.text_distributed_group =               '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[16]/a'
+        self.text_EM_chargingOrDischarging_status = '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[17]/a'
+        self.text_EM_motor_type =                   '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[18]/a'
+        self.text_EM_Parameter =                    '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[19]/a'
+        self.text_EM_Parameter_of_randomResult =    '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[20]/a'
+        self.text_EM_Pole =                         '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[21]/a'
+        self.text_EM_user_number =                  '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[22]/a'
+        self.text_EM_user_result =                  '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[23]/a'
+        self.text_EM_wholeDay_userChargingNumber =  '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[24]/a'
+        self.text_EV_chargingOrDischarging_status = '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[25]/a'
+        self.text_EV_motor_type =                   '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[26]/a'
+        self.text_EV_Parameter =                    '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[27]/a'
+        self.text_EV_Parameter_of_randomResult =    '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[28]/a'
+        self.text_EV_Pole =                         '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[29]/a'
+        self.text_EV_user_number =                  '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[30]/a'
+        self.text_EV_user_result =                  '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[31]/a'
+        self.text_EV_wholeDay_userChargingNumber =  '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[32]/a'
+        self.text_GHEMS_control_status =            '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[33]/a'
+        self.text_GHEMS_flag =                      '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[34]/a'
+        self.text_GHEMS__history_control_status =   '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[35]/a'
+        self.text_GHEMS_real_status =               '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[36]/a'
+        self.text_GHEMS_variable =                  '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[37]/a'
+        self.text_LHEMS_comfort_level =             '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[38]/a'
+        self.text_LHEMS_control_status =            '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[39]/a'
+        self.text_LHEMS_cost =                      '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[40]/a'
+        self.text_LHEMS_dr_participation =          '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[41]/a'
+        self.text_LHEMS_flag =                      '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[42]/a'
+        self.text_LHEMS__history_control_status =   '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[43]/a'
+        self.text_LHEMS_real_status =               '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[44]/a'
+        self.text_LHEMS_ucLoad =                    '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[45]/a'
+        self.text_load_list =                       '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[46]/a'
+        self.text_price =                           '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[47]/a'
+        self.text_soalr_data =                      '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[48]/a'
+        self.text_soalr_day =                       '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[49]/a'
+        self.text_totalLoad_model =                 '//*[@id="pma_navigation_tree_content"]/ul/li[4]/div[3]/ul/li[4]/div[4]/ul/li[50]/a'
+        # import use
+        self.str_BaseParameter =                    'BaseParameter'
+        self.str_EM_Parameter =                     'EM_Parameter'
+        self.str_EM_user_number =                   'EM_user_number'
+        self.str_EM_user_result =                   'EM_user_result'
+        self.str_EV_Parameter =                     'EV_Parameter'
+        self.str_EV_user_number =                   'EV_user_number'
+        self.str_EV_user_result =                   'EV_user_result'
+        self.str_GHEMS_control_status =             'GHEMS_control_status'
+        self.str_LHEMS_control_status =             'LHEMS_control_status'
+        self.str_totalLoad_model =                  'totalLoad_model'
+        self.rowCount_backup_BaseParameter =        '//*[@id="row_tbl_1"]/td[9]'
+        self.rowCount_backup_EM_Parameter =         '//*[@id="row_tbl_2"]/td[9]'
+        self.rowCount_backup_EM_user_number =       '//*[@id="row_tbl_3"]/td[9]'
+        self.rowCount_backup_EM_user_result =       '//*[@id="row_tbl_4"]/td[9]'
+        self.rowCount_backup_EV_Parameter =         '//*[@id="row_tbl_5"]/td[9]'
+        self.rowCount_backup_EV_user_number =       '//*[@id="row_tbl_6"]/td[9]'
+        self.rowCount_backup_EV_user_result =       '//*[@id="row_tbl_7"]/td[9]'
+        self.rowCount_backup_GHEMS =                '//*[@id="row_tbl_8"]/td[9]'
+        self.rowCount_backup_LHEMS =                '//*[@id="row_tbl_9"]/td[9]'
+        self.rowCount_backup_totalLoad =            '//*[@id="row_tbl_10"]/td[9]'
+        # page import
+        self.btn_import = '//*[@id="topmenu"]/li[7]/a'
+        self.btn_upload_file = '//*[@id="input_import_file"]'
+        # page export
+        self.btn_export = '//*[@id="topmenu"]/li[6]/a'
+        self.title_text_export = '//*[@id="header"]'
+        self.select_type_export = '//*[@id="plugins"]'
+        self.type_csv_export = '//*[@id="plugins"]/option[11]'
+        # page operate
+        self.btn_operate = '//*[@id="topmenu"]/li[9]/a'
+        self.text_truncate_table ='//*[@id="truncate_tbl_anchor"]'
+        self.btn_submit_truncate = '/html/body/div[8]/div[3]/div/button[1]'
+        
+        self.submit_export = '//*[@id="buttonGo"]'
+
+        self.breadcrumb = '//*[@id="breadcrumb"]/div/div/li[2]'
+        self.btn_fiftyHousehold = '/html/body/div[2]/button[4]'
+        self.btn_sweetalert = '/html/body/div[7]/div/div[6]/button[1]'
+
+        self.GHEMS_Price = '//*[@id="priceVsLoad"]'
+        self.GHEMS_SOC = '//*[@id="SOCVsLoad"]'
+        self.GHEMS_loadModel = '//*[@id="loadModel"]'
+        self.GHEMS_EMchargingSOC = '//*[@id="EMchargingSOC"]'
+        self.GHEMS_EVchargingSOC = '//*[@id="EVchargingSOC"]'
+        self.GHEMS_table = '/html/body/table'
+        self.LHEMS_loadSum = '//*[@id="households_loadsSum"]'
+        self.auto_run = '//*[@id="auto"]'
+        self.range_bar = '//*[@id="button_household_range"]/button'
+        self.range_bar_go = '/html/body/div[9]/div/div[6]/button[1]'
+        self.backup_range_bar_go = '/html/body/div[7]/div/div[6]/button[1]'
+        self.LHEMS_household_text = '//*[@id="household_id"]'
+        self.LHEMS_table = '/html/body/div[7]/table[1]'
+        self.LHEMS_household_status = '//*[@id="each_household_status"]'
+        self.LHEMS_load1 = '//*[@id="con_0"]'
+        self.LHEMS_load2 = '//*[@id="con_1"]'
+        self.LHEMS_load3 = '//*[@id="con_2"]'
+        self.LHEMS_load4 = '//*[@id="con_3"]'
+        self.LHEMS_load5 = '//*[@id="con_4"]'
+        self.LHEMS_load6 = '//*[@id="con_5"]'
+        self.LHEMS_load7 = '//*[@id="con_6"]'
+        self.LHEMS_load8 = '//*[@id="con_7"]'
+        self.LHEMS_load9 = '//*[@id="con_8"]'
+        self.LHEMS_load10 = '//*[@id="con_9"]'
+        self.LHEMS_load11 = '//*[@id="con_10"]'
+        self.LHEMS_load12 = '//*[@id="con_11"]'
+        self.LHEMS_load13 = '//*[@id="con_12"]'
+        self.LHEMS_load14 = '//*[@id="con_13"]'
+        self.LHEMS_load15 = '//*[@id="con_14"]'
+        self.LHEMS_nextHousehold_btn = '//*[@id="button_household_next"]/button[2]'
+        # html baseParameter page
+        self.baseParameter_table = '//*[@id="flag_table"]'
+        self.baseParameter_table_SOCthresh = '//*[@id="SOCthres"]'
+        self.baseParameter_table_dr_mode = '//*[@id="dr_mode"]'
+        self.baseParameter_table_simulate_weather = '//*[@id="simulate_weather"]'
+        self.baseParameter_table_simulate_price = '//*[@id="simulate_price"]'
+        self.baseParameter_table_simulate_history_weather = '//*[@id="simulate_history_weather"]'
+        self.baseParameter_table_EV_flag = '//*[@id="ElectricVehicle"]'
+        self.baseParameter_table_EM_flag = '//*[@id="ElectricMotor"]'
